rename main function from main to prontosellmagscript
